Finish first integration test of subscriptions with delayed initialization.
You can now:
 1. Bring up SlaveApiTestNode
 2. Run "rostopic echo pub std_msgs/String chatter_in hello"
 3. Eclipse console will start printing hello

Currently does not work with multiple publishers on the same topic, and
it does not shutdown cleanly.

This latest code update add a ThreadPoolExecutor to handle to
requestTopic negotiations so that they do not occur inline
with publisherUpdates.

There is also a start of some work to abstract the coupling with
the TCPROS protocol.  In the future we will want to be able to
layer in alternate protocols more easily.

This is still a work in progress -- thankfully Eclipse makes it easy
to shuffle things around.